Refactor intelligence API into modular package
